Project Library
This is a project that basically creates a library that can have books stored in it and the user can write the title, author, ISBN code and how many copies are available of the specific book. Then once ran the program displays all of the things listed above onto the console.